Add gameplay improvements and refactor code
