+++ lib/card_generator.rb
@@ -1,7 +1,8 @@
 require './lib/card'
 
 class CardGenerator
-  attr_reader :filename, :cards
+  attr_reader :filename,
+              :cards
 
   def initialize(filename)
     @filename = filename
